Remove depot identifiers from templates
Before this change, a template required a depot name be specified. This
doesn't make any sense in the context of a template, since templates
shouldn't have any knowledge of their originating depot.

Since the template object does need to know what its originating depot
is, we needed to insert that information when we scan for local
templates